+++ src/test/suite/testdata.ts
@@ -1,4 +1,4 @@
-export const testYaml = [
+export const testYamlTasks = [
     "tasks:",
     "  mow lawn: 2",
     "  eat groceries: 1",
@@ -8,4 +8,11 @@ export const testYaml = [
     "  every other day again: 2",
     "  shop: Mondays",
     "  start XMas shopping: 12/1",
-    "  pay taxes: 11/14/2021"];
+    "  pay taxes: 11/14/2021"].join("\r\n");
+
+export const testYaml = [
+    "# todotools settings for this document",
+    "runOnOpen: True",
+    "runDaily: True",
+    "lastAutoRun: 2021-11-11T04:22:18.137Z"
+].join("\r\n");